Secure Local Administrator Group Cleanup Script with Exclusions
Used to remove local administrator privileges from all accounts on the device, except for the designated IT Admin account. This proves beneficial in scenarios where a centralized IT Admin account is necessary across all devices to oversee administrative rights, while still preserving a level of local administrative access on the machine.